Check db if there is such table before attempting to delete it.
git-svn-id: https://svn.eu.apache.org/repos/asf/incubator/cayenne/main/trunk@400741 13f79535-47bb-0310-9956-ffa450edef68

+++ cayenne/src/org/objectstyle/cayenne/map/DbGenerator.java
@@ -104,12 +104,18 @@ public class DbGenerator {
       * value may also drop existing tables. */
     public void createTables(DataMap map, boolean drop) throws SQLException {
         Statement stmt = con.createStatement();
+        DatabaseMetaData meta = con.getMetaData();
 
         try {
             // DROP TABLE
             if (drop) {
                 Iterator it = map.getDbEntitiesAsList().iterator();
                 while (it.hasNext()) {
+                	ResultSet rs = meta.getTables(null, null
+                					  , ((DbEntity) it.next()).getName(), null);
+                	// If no such table, don't try to delete, just continue
+                	if (!rs.next())
+                		continue;
                     String q = adapter.dropTable((DbEntity) it.next());
                     QueryLogger.logQuery(Level.INFO, q, null);
                     stmt.execute(q);
